Extracting forms related code into a brand new concept, forms platforms
