+++ evalytics/models.py
@@ -145,3 +145,36 @@ class ReviewerResponse:
         self.eval_response = eval_response
         self.filename = filename
         self.line_number = line_number
+
+class ReviewerResponseBuilder:
+
+    def build(self, questions, filename, eval_kind, line, line_number):
+        reviewer = self.__get_reviewer_from_response_line(line)
+        reviewee = self.__get_reviewee_from_response_line(
+            line, eval_kind)
+        eval_response = self.__get_eval_response_from_response_line(
+            line, questions)
+
+        return ReviewerResponse(
+            reviewee=reviewee,
+            reviewer=reviewer,
+            eval_kind=eval_kind,
+            eval_response=eval_response,
+            filename=filename,
+            line_number=line_number
+        )
+
+    def __get_reviewer_from_response_line(self, line):
+        return line[1].split('@')[0]
+
+    def __get_reviewee_from_response_line(self, line, eval_kind):
+        if eval_kind == EvalKind.SELF:
+            reviewee = self.__get_reviewer_from_response_line(line)
+        else:
+            reviewee = line[2].strip().lower()
+
+        return reviewee
+
+    def __get_eval_response_from_response_line(self, line, questions):
+        eval_responses = line[3:]
+        return list(zip(questions, eval_responses))
